Non-recursive bit vector based implementation
- Using the "natural" combinatorial bitwise pattern to generate
permutations. A 64bit unsigned int is used to represent the project's
"presence" in the permutation, so that simple arithmetic increment is
used to traverse the permutation set the workload/value are calculated
using bitwise operations, following this logic:
    +-------------------+
    | 0 | 0 | 1 | 0 | 1 |
    +-------------------+
binary representation of decimal 5 means project[0] (lsb bit 0) and
project[2] (3rd less significant bit) are the only projects in this
"permutation" example.
- The current implementation has a limitation of at most 64 projects,
due to the undelying integer type used to implement BitSet logic
(stdint`s uint64_t).